Chore: Refactor deploy script and clean up tests (#122)
* Add prettier script and update readme

* Refactor configuration

* Organize setup for all tests

* Remove fee share from config

* Clean up imports

* More clean up

* Check chain id before setting deployed contracts

* Update readme

* Refactor and clean up tests

* Organize structure of deploy script

+++ src/interfaces/IFxGenArt721.sol
@@ -171,7 +171,7 @@ interface IFxGenArt721 is ISeedConsumer {
     function initialize(
         address _owner,
         address _primaryReceiver,
-        uint128 _lockTime,
+        uint256 _lockTime,
         ProjectInfo calldata _projectInfo,
         MetadataInfo calldata _metadataInfo,
         MintInfo[] calldata _mintInfo,
